fix(channels): pin LSP commit_fee at 1000 sat/kvB to match channel_init (#305)
* fix(channels): pin LSP commit_fee at 1000 sat/kvB to match channel_init

lsp_channels_init was reading mgr->fee (live estimatesmartfee on
testnet4/mainnet) to compute the initial commit_fee deducted from
funding_amount before splitting into local/remote.  But channel_init at
src/channel.c:282 hardcodes ch->fee_rate_sat_per_kvb=1000, and the
client mirror (client_check_conservation, src/client.c:79) derives its
expected balance from that stored 1000 rate -> base_commit_fee=154 sats.

Under testnet4 estimatesmartfee returning ~2562 sat/kvB, the LSP
actually deducted 395 sats but the client recomputed only 154 -> a
constant -241 sat conservation gap on every demo HTLC operation.

When PS_ADVANCE re-derived signing inputs from these diverged channel
states, the multi-signer quorum collapsed (only 2 of 5 nonces clean)
and client 0 EOF-disconnected.  Concretely: 3e V3b on testnet4 produced
:

    CLIENT CONSERVATION VIOLATION (after add_htlc):
      local=54303 remote=55259 htlc_sum=1043 total=110605
      expected=110846 ... base_commit_fee=154 (delta=-241)

    LSP: expected LEAF_ADVANCE_PSIG from client 0,
         got 0x00 (recv timeout / peer EOF)
    PS ADVANCE TEST: lsp_advance_leaf failed for leaf 0

Pinning the LSP-side commit_fee derivation at 1000 sat/kvB restores
agreement.  The real fee-bumping happens via CPFP on the factory tree
at broadcast time, so this initial accounting deduction does not need
to track live rates.

The existing test_fee_estimator_wiring was already setting up a
2000 sat/kvB estimator but only checked that ch->fee_rate_sat_per_kvb
stayed at 1000.  Extended it to also assert the conservation invariant
(local + remote == funding - 154) and to call lsp_channels_check_conservation
directly.  All 1469/1469 tests pass.

Regression hidden because regtest estimatesmartfee returns the 1000-sat
floor, so the bug only fired against real estimators (testnet4 /
mainnet / api / blocks modes).

* fix(htlc_commit): cast CHANNEL_DUST_LIMIT_SATS to (unsigned int) for cppcheck

PR #299 (htlc dust-race fix) introduced two fprintf calls in
htlc_commit_recv_update_fee that pass CHANNEL_DUST_LIMIT_SATS (signed
int, defined in channel.h) to a %u format specifier.  Cppcheck flagged
this as invalidPrintfArgType_uint in the static-analysis CI for #305
(stochastic — earlier PRs slipped through).

Cast to (unsigned int) to satisfy the format specifier and unblock CI.
Behavior unchanged (dust limit value is small + positive).

* fix(tests): repair 3 cheat-engine regtest scripts after defense-path shifts

Three regtest scripts started failing after today's merges
(#293/#296/#298) despite the underlying defense paths working
correctly. Root causes were all test-side:

cheat_client.sh
  Asserted "broadcast_log.source = 'penalty'" exactly, but
  PR #298 shifted WT defense for stale-leaf broadcasts onto
  the factory_response / factory_burn / subfactory_poison
  paths in watchtower.c. The strict literal filter rejected
  all defenses. Broadened to all six legitimate defense
  sources (penalty, factory_response, factory_burn,
  subfactory_poison, htlc_penalty, ptlc_penalty).

cheat_leaf_multistate.sh
  PR #296 added three new pass-signals, two of which target
  non-existent runtime artifacts:
  - "watchtower_check returned: [1-9]" -- this log line is
    never emitted by lsp_channels.c or watchtower.c (verified
    by grep across src/).  Replaced with real WT-broadcast
    markers ("Penalty tx broadcast", "Sub-factory poison tx
    broadcast", etc.).
  - "breach_detections.response_action" -- column does not
    exist in the schema (real schema is id/timestamp/
    channel_id/expected_commit_num/txid_seen/height_seen/
    response_txid, persist.c:179). Dropped the WHERE clause.
  - broadcast_log source filter was too narrow; broadened to
    all defense families (factory_/htlc_/ptlc_/poison/penalty).

cheat_daemon_leaf_multistate.sh
  Two issues:
  1) --lsp-balance-pct 50 (changed in #296 from 100) lets the
     LSP-side Tier B advance the chain before the cheater can
     successfully broadcast.  Standalone WT then has nothing
     to catch.  Reverted to 100 -- but that alone is not
     sufficient because Tier B can still win the race in
     pathological mining patterns.
  2) The wait-for-CHEAT-DAEMON-COMPLETE loop did not have an
     alternate-success exit.  When Tier B neutralizes the
     stale broadcast before the cheat code path emits the
     COMPLETE marker, the script timed out and failed.
     Added an "EARLY_TIER_B" detection path: if the LSP log
     shows "Tier B made stale state unspendable", that is a
     valid defense outcome and the script declares PASS
     directly.  Same outcome treatment in the final-result
     block when WT_FIRED is 0 but Tier B is observed.
  Also fixed a diagnostic SQL query that referenced the
  non-existent response_action column.

All 3 scripts now PASS against build-fix.  Other 25 regtest
scripts in the sweep stay green.

---------

+++ src/htlc_commit.c
@@ -369,7 +369,7 @@ int htlc_commit_recv_update_fee(channel_t *ch,
                     feerate, i,
                     (unsigned long long)ch->htlcs[i].amount_sats,
                     (unsigned long long)htlc_sweep_fee,
-                    CHANNEL_DUST_LIMIT_SATS);
+                    (unsigned int)CHANNEL_DUST_LIMIT_SATS);
             return 0;
         }
     }
@@ -384,7 +384,7 @@ int htlc_commit_recv_update_fee(channel_t *ch,
                     feerate, i,
                     (unsigned long long)ch->ptlcs[i].amount_sats,
                     (unsigned long long)htlc_sweep_fee,
-                    CHANNEL_DUST_LIMIT_SATS);
+                    (unsigned int)CHANNEL_DUST_LIMIT_SATS);
             return 0;
         }
     }
